View::update takes a `Cow<'_, Update>` instead of a `&Update`
This allows the runtime to provide an owned value if possible and a
borrowed value if not. The implementation of update can decide if
cloning is required, so we don't have to do it if the value will go
unused.

This goes especially nicely with the fact that the updates will need to
go over the wire so having an owned variant is really nice