Make `game_type` be passed as a `String` in the network protocol
This simplifies the types greatly because it means that none of the
control messages need to know about \`SupportedGames\` or \`Encoder\`.
It's a big goal to make the client not have to use the
`supported_games!` macro that the server is going to need to use. I'd
like people to be able to pull a crate off the shelf and have it "just
work" in their clients.

Just in general the goal should be for amount of work

ltti interactive players/watchers << AI builders << Game Implementers <<
Server Operators << Grant