Introduce `Wrapper` types for bots
I think some games will follow the pattern of having "phases" of play.
The `Bot::run` interface is pretty low level so it would be nice if we
could provide higher level traits that define a more game specific
interface.

The example in this commit is the `FishFightBot` trait. It defines two
seperate methods for `setup_board` and `guess`. It also removes the need
to setup the `type Game = FishFight` and wrapping returns in
`FishFight::Action`.

The other thing that I think this interface does well is that it creates
seperate views into the game depending on the "phase" the game is in.
`FishFightGuessPov` is specific to the game whereas `PlayerPov` is a
little harder to deal with directly