`Score` doesn't need `Cow`
I mean, and I really saving _that_ much?? Rankings and scores are going
to be small anyway. It's also pretty unlikely that a game will be
storing rankings and scores in a way that it can be returned as a
reference. This function is likely being called on the client anyway, so
we're not gaining much from optimizing it