Reconfigure how `BuiltinGameModes` work
The idea here is that an implentation of `LttSettings` can provide
`&'static Self`s that are keyed by a name. This means that all the games
playing one of the builtin modes can share a single instance of the
settings under the hood. This also has the advantage of making
serializing and deserializing the builtin much nicer