Depend on `semver`
I'm not totally sure this will end up staying. I haven't quite worked
out how it would work with clients with varying versions of the builtin
settings. It seems like you should be able to force the settings into a
`SettingsPtr::Custom` variant when it detects that the client isn't on a
new enough version to have that builtin. The other option is to always
transform the `SettingsPtr` to the custom variant when serializing it,
but at that point we should probably have a way of signaling to the
client that they probably have the builtin version of it