Make the `Bot::run` interface take a `&mut self`
Still feel a little uneasy about this one.

I think it's really important to be able to use information from earlier
in the game in order to affect what your bot does this turn, but the
question is how do we want to provide that data to the bot.

It seems weird for the bot to need to keep track of the history of the
game when `ltt` provides state machines for all the major components of
the game, but it seems weird for things like `GamePlayer` to be in
charge in providing views into historical data, _especially_ if a bot
decides it doesn't need that info.

Then comes the question is there a generic thing that a bot could do to
"opt in" to history, even if we just provide a container to store it's
updates in. It also means we likely need callbacks even when the bot
isn't called to act, passing the update itself into the bot so it can
change it's state based on the data it just got. The whole thing feels
pretty weird and I need to think on it more

The other open question is if there is a way to give a "Crash Report" to
a player with the contents of the bot state, along with the game player
state that cause it to fail. It would be really cool to have a "create
unit test from my last crash" feature in stadium