A very basic public info action flow for the game
I'm trying to balance how much info to pass through the `Phase` field of
the `PublicInfo` it's tricky because any logic that gets written into
`BlackJackBot` needs to be replicated in any other bot implementation.
It would be nice to have a dead simple `Bot` implementation for anything
wanting to play black jack. The other half of that is wanting to avoid
denormalizing data. It would be really nice if I could avoid having to
make sure multiple sources of data are in sync